+++ tardsquad_discord_bot/commands_rating.py
@@ -4,14 +4,14 @@ from operator import attrgetter
 from discord.ext import commands
 
 import tardsquad_discord_bot
+from tardsquad_discord_bot.cog import TardBotCog
 from tardsquad_discord_bot.ratingentry import RatingEntry
 from tardsquad_discord_bot.textcolor import TextColor
 
 
-class TardBotRatingCommands(commands.Cog, name="Rating"):
-    def __init__(self, bot, guild, *args, **kwargs):
-        self.bot = bot
-        self.guild = guild
+class TardBotRatingCommands(TardBotCog, name="Rating"):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
         self.ratings = {}
 
     def _rating_gc(self):