+++ tardsquad_discord_bot/commands_search.py
@@ -3,14 +3,11 @@ import urllib
 from discord.ext import commands
 
 import tardsquad_discord_bot
+from tardsquad_discord_bot.cog import TardBotCog
 from tardsquad_discord_bot.textcolor import TextColor
 
 
-class TardBotSearchCommands(commands.Cog, name="Search"):
-    def __init__(self, bot, guild, *args, **kwargs):
-        self.bot = bot
-        self.guild = guild
-
+class TardBotSearchCommands(TardBotCog, name="Search"):
     @staticmethod
     def search_query(url_fmt, *args):
         query = urllib.parse.quote(" ".join(args))