+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.9.0] - 2021-12-28
 ### Changed
 - Downgrade `!rate 10` to `9` if not entitled.
 - Rename `!rate-all` to `!rate-standup`.